libsframe: use uint8_t for return type of sframe_fre_get_base_reg_id
Use a more appropriate data type.

include/
	* sframe-api.h (sframe_fre_get_base_reg_id): Use uint8_t as
	return type.
libsframe/
	* sframe-dump.c (dump_sframe_func_with_fres): Use uint8_t type
	for base reg id.
	* sframe.c (sframe_fre_get_base_reg_id): Use uin8_t as return
	type.

+++ libsframe/sframe.c
@@ -618,7 +618,7 @@ sframe_calc_fre_type (size_t func_size)
 
 /* Get the base reg id from the FRE info.  Set errp if failure.  */
 
-unsigned int
+uint8_t
 sframe_fre_get_base_reg_id (sframe_frame_row_entry *fre, int *errp)
 {
   if (fre == NULL)
